Visual and Directories
Visual updates to the borders.
Major directory structure overhaul to the dreamcast website half. Plan
to expand that to the PC side. This should simplify navigation and
referencing directories. Rather than working your way back to specific
points to get to the image directory, you specify the image directory
and it already includes all necessary navigation in that variable.